+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
 ## Version 0.23.0
+* Fix ability to grant whitelist to all other roles as long as grantee has AUTHORIZE permission on the resource (Ericsson/ecaudit#31)
+  * __NOTE__: As of this version it is required to configure the AuditAuthorizer in the cassandra.yaml - refer to the [setup guide](doc/setup.md) for details
 * Improve logger performance with micro benchmarks
 
 ## Version 0.22.0